requirements.txt
bson==0.5.8
coverage==4.5.4
cryptography==2.7.0
flake8==3.7.2
Flask==1.1.1
flask-cors==3.0.8
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
logzero==1.5.0
marshmallow==2.20.2
marshmallow-sqlalchemy==0.17.0
pre-commit==1.18.2
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0